+++ msgpack_helper_test.go
@@ -4,6 +4,7 @@
 package tarantool_test
 
 import (
+	"github.com/tarantool/go-tarantool"
 	"gopkg.in/vmihailenco/msgpack.v2"
 )
 
@@ -12,4 +13,17 @@ type decoder = msgpack.Decoder
 
 func encodeUint(e *encoder, v uint64) error {
 	return e.EncodeUint(uint(v))
+}
+
+func toBoxError(i interface{}) (v tarantool.BoxError, ok bool) {
+	v, ok = i.(tarantool.BoxError)
+	return
+}
+
+func marshal(v interface{}) ([]byte, error) {
+	return msgpack.Marshal(v)
+}
+
+func unmarshal(data []byte, v interface{}) error {
+	return msgpack.Unmarshal(data, v)
 }