+++ msgpack_v5_helper_test.go
@@ -4,6 +4,7 @@
 package tarantool_test
 
 import (
+	"github.com/tarantool/go-tarantool"
 	"github.com/vmihailenco/msgpack/v5"
 )
 
@@ -12,4 +13,20 @@ type decoder = msgpack.Decoder
 
 func encodeUint(e *encoder, v uint64) error {
 	return e.EncodeUint(v)
+}
+
+func toBoxError(i interface{}) (v tarantool.BoxError, ok bool) {
+	var ptr *tarantool.BoxError
+	if ptr, ok = i.(*tarantool.BoxError); ok {
+		v = *ptr
+	}
+	return
+}
+
+func marshal(v interface{}) ([]byte, error) {
+	return msgpack.Marshal(v)
+}
+
+func unmarshal(data []byte, v interface{}) error {
+	return msgpack.Unmarshal(data, v)
 }